brlock: Always reference the argument
The brlock primitives take a kvm object as argument but don't always
reference it. This can lead to unused variable warnings (treated as
error) when compiling.

Always make a reference to the argument. This also ensures that the
argument expression is always evaluated, in case in the future some
nasty mind relies on possible side effects of the expression.

Signed-off-by: Julien Thierry <julien.thierry@arm.com>